requirements.txt
# Common requirements
Django~=2.2.10
django-cors-headers~=3.1.1
django-extensions~=2.2.8
django-filter~=2.2.0
djangorestframework~=3.11
djangorestframework-queryfields~=1.0.0
dry-rest-permissions~=0.1.10
drf_yasg~=1.17.1
mysqlclient~=1.4
python-memcached~=1.59